+++ migrator/src/database/left_db.rs
@@ -1,0 +1,228 @@
+/*
+ * Author: Oyelowo Oyedayo
+ * Email: oyelowo.oss@gmail.com
+ * Copyright (c) 2023 Oyelowo Oyedayo
+ * Licensed under the MIT license
+ */
+
+use std::ops::Deref;
+
+use surreal_orm::{
+    statements::{begin_transaction, delete, select_value},
+    *,
+};
+use surrealdb::{Connection, Surreal};
+
+use crate::*;
+
+use super::{
+    file_manager::MigrationFileName,
+    migration_meta::{Migration, MigrationBidirectional, MigrationUnidirectional, Mode},
+    migrator_db::{LeftFullDbInfo, MigratorDatabase},
+};
+
+// For the migration directory
+#[derive(Debug, Clone)]
+pub struct LeftDatabase(pub MigratorDatabase);
+
+impl Deref for LeftDatabase {
+    type Target = MigratorDatabase;
+
+    fn deref(&self) -> &Self::Target {
+        &self.0
+    }
+}
+
+impl LeftDatabase {
+    pub async fn resources(&self) -> LeftFullDbInfo {
+        LeftFullDbInfo(
+            self.0
+                .get_all_resources()
+                .await
+                .expect("nothing for u on left"),
+        )
+    }
+
+    pub async fn run_local_dir_up_migrations<C: Connection>(
+        db: Surreal<C>,
+        migrations: Vec<MigrationBidirectional>,
+    ) -> MigrationResult<()> {
+        let queries = migrations
+            .iter()
+            .map(|m| m.up.clone())
+            .collect::<Vec<_>>()
+            .join("\n");
+
+        let marked_up_migrations = migrations
+            .iter()
+            .map(|m| {
+                Migration {
+                    id: Migration::create_id(m.id.clone().to_string()),
+                    name: m.name.clone(),
+                    timestamp: m.timestamp,
+                }
+                .create()
+                .to_raw()
+                .build()
+            })
+            .collect::<Vec<_>>()
+            .join("\n");
+
+        let all = format!("{}\n{}", queries, marked_up_migrations);
+
+        // Run them as a transaction against a local in-memory database
+        if !queries.trim().is_empty() {
+            begin_transaction()
+                .query(Raw::new(all))
+                .commit_transaction()
+                .run(db)
+                .await?;
+        }
+        Ok(())
+    }
+
+    pub async fn run_all_local_dir_up_migrations(&self, mode: Mode) -> MigrationResult<()> {
+        let all_migrations = MigrationBidirectional::get_all_from_migrations_dir(mode)?;
+        Self::run_local_dir_up_migrations(self.db(), all_migrations).await?;
+        Ok(())
+    }
+
+    pub async fn run_local_dir_oneway_content_migrations<C: Connection>(
+        db: Surreal<C>,
+        migrations: Vec<MigrationUnidirectional>,
+    ) -> MigrationResult<()> {
+        let queries = migrations
+            .iter()
+            .map(|m| m.content.clone())
+            .collect::<Vec<_>>()
+            .join("\n");
+        println!("Running queries: {}", queries);
+
+        let marked_up_migrations = migrations
+            .iter()
+            .map(|m| {
+                Migration {
+                    id: Migration::create_id(m.id.clone().to_string()),
+                    name: m.name.clone(),
+                    timestamp: m.timestamp,
+                }
+                .create()
+                .to_raw()
+                .to_string()
+            })
+            .collect::<Vec<_>>()
+            .join("\n");
+
+        let all = format!("{}\n{}", queries, marked_up_migrations);
+
+        // Run them as a transaction against a local in-memory database
+        if !queries.trim().is_empty() {
+            begin_transaction()
+                .query(Raw::new(all))
+                .commit_transaction()
+                .run(db)
+                .await?;
+        }
+        Ok(())
+    }
+    pub async fn run_all_local_dir_one_way_migrations(&self, mode: Mode) -> MigrationResult<&Self> {
+        let all_migrations = MigrationUnidirectional::get_all_from_migrations_dir(mode)?;
+        let queries = all_migrations
+            .into_iter()
+            .map(|m| m.content)
+            .collect::<Vec<_>>()
+            .join("\n");
+
+        // Run them as a transaction against a local in-memory database
+        if !queries.trim().is_empty() {
+            begin_transaction()
+                .query(Raw::new(queries))
+                .commit_transaction()
+                .run(self.db())
+                .await?;
+        }
+        Ok(self)
+    }
+
+    pub async fn get_applied_bi_migrations_from_db(&self) -> MigrationResult<Vec<Migration>> {
+        // let name = Migration::schema().name;
+        let migration::Schema { name, .. } = Migration::schema();
+        let migration = Migration::table_name();
+
+        // select [{ name: "Oyelowo" }]
+        // select value [ "Oyelowo" ]
+        // select_only. Just on object => { name: "Oyelowo" }
+        let migration_names = select_value(name)
+            .from(migration)
+            .return_many::<Migration>(self.db())
+            .await?;
+        Ok(migration_names)
+    }
+
+    // pub async fn get_applied_uni_migrations_from_db(&self) -> MigrationResult<Vec<String>> {
+    //     let migration_unidirectional::Schema { name, .. } = MigrationUnidirectional::schema();
+    //     let migration = MigrationUnidirectional::table_name();
+    //
+    //     // select [{ name: "Oyelowo" }]
+    //     // select value [ "Oyelowo" ]
+    //     // select_only. Just on object => { name: "Oyelowo" }
+    //     let migration_names = select_value(name)
+    //         .from(migration)
+    //         .return_many::<String>(self.db())
+    //         .await?;
+    //     Ok(migration_names)
+    // }
+
+    pub async fn mark_migration_as_applied(
+        &self,
+        migration_name: impl Into<MigrationFileName>,
+    ) -> MigrationResult<Migration> {
+        let migration_name: MigrationFileName = migration_name.into();
+        println!("Applying migration: {}", migration_name);
+
+        let migration = Migration {
+            id: Migration::create_id(migration_name.to_string()),
+            name: migration_name.to_string(),
+            timestamp: migration_name.timestamp(),
+        }
+        .create()
+        .get_one(self.db())
+        .await?;
+        println!("Migration applied: {}", migration_name);
+
+        Ok(migration)
+    }
+
+    pub async fn unmark_migration(&self, migration_name: MigrationFileName) -> MigrationResult<()> {
+        println!("Unmark migration: {}", migration_name);
+        delete::<Migration>(Migration::create_id(migration_name.to_string()))
+            .run(self.db())
+            .await?;
+        println!("Migration unmarked: {}", migration_name);
+        Ok(())
+    }
+
+    pub async fn rollback_migration(
+        db: &mut Self,
+        migration_name: MigrationFileName,
+    ) -> MigrationResult<()> {
+        let migration =
+            MigrationBidirectional::get_migration_by_name(migration_name.clone(), db.mode)?;
+        if let Some(migration) = migration {
+            let down_migration = migration.down;
+            if !down_migration.trim().is_empty() {
+                // Raw::new(down_migration).run(db);
+                db.execute(down_migration).await?;
+            } else {
+                println!("No down migration found for migration: {}", migration_name);
+            }
+            db.unmark_migration(migration.name.try_into()?).await?;
+        } else {
+            println!(
+                "Cannot rollback migration: No migration found with name: {}",
+                migration_name
+            );
+        };
+        Ok(())
+    }
+}
