+++ migrator/src/database/mod.rs
@@ -1,0 +1,22 @@
+/*
+ * Author: Oyelowo Oyedayo
+ * Email: oyelowo.oss@gmail.com
+ * Copyright (c) 2023 Oyelowo Oyedayo
+ * Licensed under the MIT license
+ */
+
+mod file_manager;
+mod full_db_info;
+mod left_db;
+mod migration_meta;
+mod migrator_db;
+mod queries;
+mod right_db;
+
+pub use file_manager::*;
+pub use full_db_info::*;
+pub use left_db::*;
+pub use migration_meta::*;
+pub use migrator_db::*;
+pub use queries::*;
+pub use right_db::*;
