+++ migrator/src/models/mod.rs
@@ -1,0 +1,117 @@
+/*
+ * Author: Oyelowo Oyedayo
+ * Email: oyelowo.oss@gmail.com
+ * Copyright (c) 2023 Oyelowo Oyedayo
+ * Licensed under the MIT license
+ */
+
+use chrono::Utc;
+use serde::{Deserialize, Serialize};
+use surreal_orm::{
+    arr, cond, create_table_resources,
+    statements::{define_event, define_index, select},
+    *,
+};
+
+#[derive(Node, Serialize, Deserialize, Debug, Clone, Default)]
+#[serde(rename_all = "camelCase")]
+#[surreal_orm(table_name = "planet", schemafull)]
+pub struct Planet {
+    // Test renaming tomorrow
+    pub id: SurrealSimpleId<Self>,
+    pub first_name: String,
+    pub population: u64,
+    pub created_at: chrono::DateTime<Utc>,
+    pub tags: Vec<String>,
+}
+
+impl TableResources for Planet {}
+
+#[derive(Node, Serialize, Deserialize, Debug, Clone, Default)]
+#[serde(rename_all = "camelCase")]
+#[surreal_orm(table_name = "student", schemafull)]
+pub struct Student {
+    pub id: SurrealSimpleId<Self>,
+    pub high_school: String,
+    pub age: u8,
+}
+
+impl TableResources for Student {}
+
+#[derive(Node, Serialize, Deserialize, Debug, Clone, Default)]
+#[serde(rename_all = "camelCase")]
+#[surreal_orm(table_name = "animal", schemafull)]
+pub struct Animal {
+    pub id: SurrealSimpleId<Self>,
+    pub species: String,
+    // Improve error essage for old_nmae using word similarity algo
+    pub attributes: Vec<String>,
+    pub updated_at: chrono::DateTime<Utc>,
+    pub velocity: u64,
+}
+
+impl TableResources for Animal {
+    fn events_definitions() -> Vec<Raw> {
+        let animal::Schema {
+            species, velocity, ..
+        } = Self::schema();
+
+        let event1 = define_event("event1".to_string())
+            .on_table("animal".to_string())
+            .when(cond(species.eq("Homo Erectus").and(velocity.gt(545))))
+            .then(select(All).from(Crop::table_name()))
+            .to_raw();
+
+        let event2 = define_event("event2".to_string())
+            .on_table("animal".to_string())
+            .when(cond(species.eq("Homo Sapien")).and(velocity.lt(10)))
+            .then(select(All).from(AnimalEatsCrop::table_name()))
+            .to_raw();
+        vec![event1, event2]
+    }
+
+    fn indexes_definitions() -> Vec<Raw> {
+        let animal::Schema {
+            species, velocity, ..
+        } = Self::schema();
+
+        let idx1 = define_index("species_speed_idx".to_string())
+            .on_table(Self::table_name())
+            .fields(arr![species, velocity])
+            .unique()
+            .to_raw();
+
+        vec![idx1]
+    }
+}
+
+pub struct Resources;
+
+impl DbResources for Resources {
+    create_table_resources!(Animal, Crop, AnimalEatsCrop, Student, Planet);
+}
+
+#[derive(Edge, Serialize, Deserialize, Debug, Clone, Default)]
+#[serde(rename_all = "camelCase")]
+#[surreal_orm(table_name = "eats", schemafull)]
+pub struct Eats<In: Node, Out: Node> {
+    pub id: SurrealSimpleId<Self>,
+    #[serde(rename = "in")]
+    pub in_: In,
+    pub out: Out,
+    pub place: String,
+    pub created_at: chrono::DateTime<Utc>,
+}
+
+pub type AnimalEatsCrop = Eats<Animal, Crop>;
+impl TableResources for AnimalEatsCrop {}
+
+#[derive(Node, Serialize, Deserialize, Debug, Clone, Default)]
+#[serde(rename_all = "camelCase")]
+#[surreal_orm(table_name = "crop", schemafull)]
+pub struct Crop {
+    pub id: SurrealSimpleId<Self>,
+    pub color: String,
+}
+
+impl TableResources for Crop {}
