+++ query-builder/src/traits/mod.rs
@@ -6,6 +6,7 @@
  */
 
 pub(crate) mod binding;
+pub(crate) mod db_resources;
 pub(crate) mod general;
 pub(crate) mod model;
 pub(crate) mod operation;
@@ -13,8 +14,10 @@ pub(crate) mod patch_op;
 pub(crate) mod raw;
 pub(crate) mod setter;
 pub(crate) mod statements;
+pub(crate) mod table_resources;
 
 pub use binding::*;
+pub use db_resources::*;
 pub use general::*;
 pub use model::*;
 pub use operation::*;
@@ -22,3 +25,4 @@ pub use patch_op::*;
 pub use raw::*;
 pub use setter::*;
 pub use statements::*;
+pub use table_resources::*;
